Silence a compilation warning on Windows
WL#7170 InnoDB buffer estimates for tables and indexes

+++ storage/innobase/include/ut0lock_free_hash.h
@@ -142,7 +142,8 @@ public:
 			void*		mem;
 
 			if (m_numa_available) {
-				const int	node = os_numa_node_of_cpu(i);
+				const int	node = os_numa_node_of_cpu(
+					static_cast<int>(i));
 
 				mem = os_numa_alloc_onnode(s, node);
 			} else {
